requirements.txt
numpy
opencv-python
opencv-contrib-python
imutils
matplotlib
pillow
scikit-image